Move function for parsing backend token from url from Protocol to BackendToken, and reuse it in Backend code instead of relying on fragile regex.
